third_party/blink/web_tests/external/wpt/html/webappapis/scripting/processing-model-2/integration-with-the-javascript-job-queue/promise-job-detached-iframe-expected.txt

This is a testharness.js-based test.
[FAIL] Reaction job should be skipped if onFulfilled handler belongs to detached iframe
  assert_equals: expected false but got true
[FAIL] Reaction job should be skipped if onRejected handler belongs to detached iframe
  assert_equals: expected false but got true
[FAIL] Thenable job for fulfilled promise should be skipped if it belongs to detached iframe
  assert_equals: expected false but got true
[FAIL] Thenable job for rejected promise should be skipped if it belongs to detached iframe
  assert_equals: expected false but got true
[FAIL] Thenable job created by reaction job for fulfillment should be skipped if it belongs to detached iframe
  assert_equals: expected false but got true
[FAIL] Thenable job created by reaction job for rejection should be skipped if it belongs to detached iframe
  assert_equals: expected false but got true
[FAIL] Thenable job created by Promise.all should be skipped if it belongs to detached iframe
  assert_equals: expected false but got true
[FAIL] Thenable job created by Promise.allSettled should be skipped if it belongs to detached iframe
  assert_equals: expected false but got true
[FAIL] Thenable job created by Promise.any should be skipped if it belongs to detached iframe
  assert_equals: expected false but got true
[FAIL] Thenable job created by Promise.race should be skipped if it belongs to detached iframe
  assert_equals: expected false but got true
Harness: the test ran to completion.
